+++ server/data/abilities.js
@@ -1,12 +1,5 @@
-const { NotificationType } = require('../objects/Player')
+const { NotificationType, AbilityTag } = require('../data/enums')
 const { GameManager } = require('../utils/GameManager')
-
-const AbilityTag = Object.freeze({
-    DAY: 0, // use only during day
-    NIGHT: 1, // use only during night
-    DESIGNATED: 2, // need DA status to use it
-    ASTRAL: 3 // does not visit
-})
 
 const abilityDictionary = {
     "PLACEHOLDER_INVESTIGATE": {
@@ -53,4 +46,4 @@ const abilityDictionary = {
     }
 }
 
-module.exports = { abilityDictionary, AbilityTag }
+module.exports = { abilityDictionary }
