+++ server/utils/AbilityManager.js
@@ -1,5 +1,5 @@
+const { AbilityTag } = require('../data/enums')
 const { GameManager } = require("./GameManager")
-const { AbilityTag } = require('../data/abilities')
 
 class AbilityManager {
     /** @type {Array<{user: string, id: string, targetData: Array}>} */
